requirements.txt
# Core dependencies
python-dotenv>=1.0.0
pydantic>=2.5.0

# Document parsing
PyPDF2>=3.0.0
python-docx>=1.0.0
python-pptx>=0.6.21
pandas>=2.0.0
markdown>=3.5.0

# Vector store and embeddings
chromadb>=0.4.0
sentence-transformers>=2.2.0
openai>=1.0.0

# Web interface
streamlit>=1.29.0

# Utilities
python-multipart>=0.0.6
typing-extensions>=4.8.0
loguru>=0.7.0
